Fix windows / msvc build warnings (#42)
* Fix windows / msvc build warnings - signed/unsigned mismatch
```
D:\a\1\s\libsrc\TextBox.cc(167): warning C4018: '<': signed/unsigned mismatch [D:\a\1\a\mylib.vcxproj]
D:\a\1\s\libsrc\TextBox.cc(189): warning C4018: '<': signed/unsigned mismatch [D:\a\1\a\mylib.vcxproj]
```

* Fix windows build warnings - `asctime`
```
D:\a\1\s\libsrc\File.cc(77): warning C4996: 'asctime': This function or variable may be unsafe. Consider using asctime_s instead.
To disable deprecation, use _CRT_SECURE_NO_WARNINGS.
See online help for details. [D:\a\1\a\mylib.vcxproj]
  C:\Program Files (x86)\Windows Kits\10\Include\10.0.14393.0\ucrt\time.h(116): note: see declaration of 'asctime'
```

* Fix windows build warnings - `localtime`
```
D:\a\1\s\libsrc\File.cc(78): warning C4996: 'localtime': This function or variable may be unsafe. Consider using localtime_s instead.
To disable deprecation, use _CRT_SECURE_NO_WARNINGS.
See online help for details. [D:\a\1\a\mylib.vcxproj]
  C:\Program Files (x86)\Windows Kits\10\Include\10.0.14393.0\ucrt\time.h(505): note: see declaration of 'localtime'
```

+++ libsrc/TextBox.h
@@ -46,7 +46,7 @@ public:
   };
 protected:
   char *text;
-  int num_lines;
+  size_t num_lines;
   char box_char;
   box_style box_type;
 public:
